Almost Finished
The Radial Slider is now added, replacing the normal slider used before.
Adding a new sample gives you the option to give it a name.
The name of each sample is now displayed next to its respective row.
Edited how the Play and Stop buttons look.
Removed the Pause button.